Trait interrupt_support::Interruptee

pub trait Interruptee {
    fn was_interrupted(&self) -> bool;

    fn err_if_interrupted(&self) -> Result<(), Interrupted> { ... }
}

Represents the state of something that may be interrupted. Decoupled from the interrupt mechanics so that things which want to check if they have been interrupted are simpler.

Required methods

fn was_interrupted(&self) -> bool

Loading content...
Provided methods

fn err_if_interrupted(&self) -> Result<(), Interrupted>

Loading content...