Crate interrupt_support

Structs

Interrupted

The error returned by err_if_interrupted.

NeverInterrupts

A convenience implementation, should only be used in tests.

ShutdownInterruptee

SqlInterruptHandle

Interrupt operations that use SQL

SqlInterruptScope

Check if an operation has been interrupted

Traits

Interruptee

Represents the state of something that may be interrupted. Decoupled from the interrupt mechanics so that things which want to check if they have been interrupted are simpler.

Functions

in_shutdown

Check if we’re currently in shutdown mode

register_interrupt

Register a ShutdownInterrupt implementation

shutdown

Initiate shutdown mode